+++ packages/array-includes-with-glob/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.3.0 - 2017-04-23
+### Added
+- You can provide an array (second argument) of strings to check, is ANY of them found in the source (first argument).
+- Alternatively, you can set the matching to request that ALL elements from second argument array must be found in the source array (first argument).
+
 ## 1.2.0 - 2017-04-22
 ### Added
 - Now the input can be not only array, but string as well. It will be converted into an array first and then processed.